+++ src/services/system/permission-service/src/application/queries/role/index.ts
@@ -8,17 +8,21 @@ export * from './list-role-permissions.query'
 export * from './list-role-permissions.handler'
 export * from './list-role-accounts.query'
 export * from './list-role-accounts.handler'
+export * from './get-account-role-selection.query'
+export * from './get-account-role-selection.handler'
 
 import { GetRoleByIdHandler } from './get-role-by-id.handler'
 import { ListRolesHandler } from './list-roles.handler'
 import { ListAccountRolesHandler } from './list-account-roles.handler'
 import { ListRolePermissionsHandler } from './list-role-permissions.handler'
 import { ListRoleAccountsHandler } from './list-role-accounts.handler'
+import { GetAccountRoleSelectionHandler } from './get-account-role-selection.handler'
 
 export const RoleQueryHandlers = [
   GetRoleByIdHandler,
   ListRolesHandler,
   ListAccountRolesHandler,
   ListRolePermissionsHandler,
-  ListRoleAccountsHandler
+  ListRoleAccountsHandler,
+  GetAccountRoleSelectionHandler
 ]